requirements.txt
python-telegram-bot==20.7
gspread==5.12.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pytz==2024.1
python-dotenv==1.0.1
